+++ tests/integration/test_ast.py
@@ -1,8 +1,6 @@
-import threading
-
 import pytest
 
-from rethinkdb.ast import Repl, dict_items, needs_wrap
+from rethinkdb.ast import Repl
 
 
 @pytest.fixture
@@ -10,10 +8,11 @@ def repl():
     """
     Fixture returning a Repl object instance for testing purpose.
     """
+
     return Repl()
 
 
-def test_repl_get(repl):
+def test_repl_get(repl: Repl):  # pylint: disable=redefined-outer-name
 
-    res = repl.get()
+    res = repl.get_connection()
     assert res == None