+++ CHANGELOG.rst
@@ -33,6 +33,8 @@ Changed
 * `HandshakeV1_0` defines `username` and `password` attributes as protected attributes
 * `HandshakeV1_0` has a hardcoded `JSONEncoder` and `JSONDecoder` from now on
 * `HandshakeV1_0` raises `InvalidHandshakeStateError` when an unrecognized state called in `next_message`
+* Moved `ReQLEncoder`, `ReQLDecoder`, `recursively_make_hashable` to `encoder` module
+* Moved `T` to `utilities` to module and renamed to `EnhancedTuple`
 
 Removed
 ~~~~~~~